Allow adjusting pointer if newly compressed file is bigger than the original
This crashes in game due to the arithmetic used to calculate the address of the next file. I will address this issue in a future commit. It will probably involve recomputing the entire compressed filesystem and shifting the audio data (at the end of the ROM) further down. Quite a daunting task.